+++ src/components/OnPremise/GroupWorkBottomSheet.tsx
@@ -27,7 +27,7 @@ const GroupWorkBottomSheet = ({
         {t('onPremise.groupWork.label')}
       </AppText>
       <AppText
-        style={tw`text-left text-base font-normal text-slate-500 dark:text-neutral-500 mt-6`}>
+        style={tw`text-left text-base font-normal text-slate-500 dark:text-neutral-500 mt-6 mb-3`}>
         {t('onPremise.groupWork.description')}
       </AppText>
     </AppBottomSheet>