+++ src/components/OnPremise/WifiBottomSheet.tsx
@@ -1,13 +1,13 @@
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import * as Haptics from 'expo-haptics';
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { StyleProp, View, ViewStyle } from 'react-native';
 import Animated, { FadeIn, FadeOutDown } from 'react-native-reanimated';
 import { RandomReveal } from 'react-random-reveal';
 import tw from 'twrnc';
 import WifiNetworkAnimation from '@/components/Animations/WifiNetworkAnimation';
-import AppBottomSheet from '@/components/AppBottomSheet';
+import AppBottomSheet, { AppBottomSheetRef } from '@/components/AppBottomSheet';
 import AppRoundedButton from '@/components/AppRoundedButton';
 import AppText from '@/components/AppText';
 import AppTextLink from '@/components/AppTextLink';
@@ -26,6 +26,7 @@ const WifiBottomSheet = ({
   onClose?: () => void;
 }) => {
   const { t } = useTranslation();
+  const bottomSheetRef = useRef<AppBottomSheetRef>(null);
   const user = useAuthStore((s) => s.user);
   const noticeStore = useNoticeStore();
   const [ssid, setSSID] = useState<string | null>(null);
@@ -49,6 +50,7 @@ const WifiBottomSheet = ({
 
   return (
     <AppBottomSheet
+      ref={bottomSheetRef}
       contentContainerStyle={tw`flex flex-col items-stretch pt-6 px-6`}
       style={style}
       onClose={onClose}>
@@ -62,7 +64,12 @@ const WifiBottomSheet = ({
 
       <Trans
         components={[
-          <AppTextLink href={`/devices`} key="add-device-link" style={tw`text-amber-500`} />,
+          <AppTextLink
+            href={`/devices`}
+            key="add-device-link"
+            style={tw`text-amber-500`}
+            onPress={() => bottomSheetRef.current?.close()}
+          />,
           <AppTextLink
             href={`${WORDPRESS_BASE_URL}/mon-compte/appareils`}
             key="add-device-through-website-link"
@@ -72,11 +79,11 @@ const WifiBottomSheet = ({
         ]}
         defaults={t('onPremise.wifi.description')}
         parent={AppText}
-        style={tw`text-left text-base font-normal text-slate-500 dark:text-neutral-500 mt-6`}
+        style={tw`text-left text-base font-normal text-slate-500 dark:text-neutral-500 mt-6 mb-3`}
       />
 
       {ssid || password ? (
-        <Animated.View entering={FadeIn.delay(100)} style={tw`my-6 flex flex-col`}>
+        <Animated.View entering={FadeIn.delay(100)} style={tw`my-3 flex flex-col`}>
           <SectionTitle title={t('onPremise.wifi.credentials.ssid.label')} />
           <AppText style={tw`text-left text-slate-900 dark:text-gray-200 text-2xl font-bold`}>
             {ssid && <RandomReveal isPlaying characters={ssid} duration={2} />}
@@ -88,7 +95,7 @@ const WifiBottomSheet = ({
           </AppText>
         </Animated.View>
       ) : (
-        <Animated.View exiting={FadeOutDown} style={tw`w-full mt-3`}>
+        <Animated.View exiting={FadeOutDown} style={tw`w-full`}>
           <AppRoundedButton
             disabled={!user?.capabilities?.includes('WIFI_CREDENTIALS_ACCESS')}
             loading={isLoading}