+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Animate error badge icon.
+- Link to WiFi credentials bottom sheet on pairing a new device.
 
 ## [2026.1.2] - 2026-01-28